Add UI test for survey
